Fix javadoc issues in system APIs reported by doclava
Fixes some errors in system API javadoc that are reported by doclava.

This was not an	issue before because doclava was ignoring all
`@SystemApi` annotated APIs because while Metalava removed the
`@SystemApi` annotations it left the `@hide`. That meant that doclava
just ignored them. When that Metalava bug was fixed these problems were
revealed.

Bug: 462333324
Test: m offline-system-sdk-referenceonly-docs
Flag: EXEMPT DOCS_ONLY
Change-Id: I5bfaca0a3e7f197deae9f2f5db47939f75b3a6c7