Fix shade getting stuck open on top of the dream
This happens because when the dream is showing, we hide the scene
container and forward the touches to it. However, we are not forwarding
the UP/CANCEL event, so the scene container doesn't know the user
stopped interacting with the device. This leads to the scene transition
getting stuck.

Fixes: 461795463
Flag: com.android.systemui.scene_container
Test: atest BouncerSwipeTouchHandlerTest
Test: atest ShadeTouchHandlerTest
Change-Id: I8c07f931aec251078e28b725be4d21b6ac964a4c