Add AiSeal test
Add AiSeal integration test with configuration files and sample tenants.

The test installs all test tenants as platform application and updates
AiSeal configuration to start the VM with test tenants. Then it uses
services exported by the tenants to check that they can connect to the
VM. Test shouldn't be able to connect to the VM directly, because it is
not a tenant of the VM itself.

AiSeal API is accessible only to privileged and platform application and
its configuration couldn't be modified without root and modifying system
partition. This is done to ensure that AiSeal isolation couldn't be
broken even with access to adb shell.

MODULE_NOT_SECONDARY_USER_REASON=AiSeal API is available only for system
user, because profile separation doesn't exist in VM and tenants have to
manage profile data separation from host side.

Bug: 455594108
Test: atest AiSealTestCases
Flag: android.aiseal.aiseal_host_apis
Change-Id: I78f870b8471cdf5a4db5e39dbd898056f3b21828