Use DelayQueue in the fake executor.
This is an alternative approach for adding thread safety. This also
simplifies some of the code because it can handle the logic for checking
if the futures have reached their execution point.

The main disadvantage is that we don't actually want blocking
operations, or anything using a timeout, so one has to be careful to
only use the right operations. But this is probably still safer than
relying on us manually doing the synchronization.

Bug: 461796483
Test: mmma, atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager
Change-Id: I100ff552e0780c8d5c4130d5f157a32ddfc37ccf