Fix AdbdMessage parsing
There is a off by one error in the parsing routine. We don't hit it in
production because we pass a very large buffer to the socket reading
from adbd_auth.

It is really weird that we did not see a failure for this test since
it is supposed to be part of pre-submit.

Test: AdbdMessageTests
Bug: 449934216
Flag: EXEMPT BUGFIX
Change-Id: I40be58ceac19e6f64ac557e3c557056a3dce47f1