Don't set task view bounds if the bubble bar position is stale
When the positioner is updated with a new configuration, the bubble
bar position changes as a result. This change waits until the new
position is sent from launcher before updating the task view bounds.

Additionally remove onLocationChanged from BubbleBarExpandedView.
This is currently called after setting the x and y coordinates for
the expanded view which trigger a layout pass and we already update
the task view bounds in onLayout. So having them one after the other
triggers 2 bounds updates (with different values).

Bug: 457332269
Flag: com.android.wm.shell.enable_bubble_bar
Test: manual
       - checked that the expanded view works as expected
Test: manual
       - create a bubble bar bubble
       - expand it
       - rotate the device such that the camera cutout is at the bottom
       - verify the expanded view shows correctly
       - rotate back and verify expanded view
Change-Id: I3d3b2cb276aab3f1df41aec8f813112061977100