[flexiglass] Remove extra transition to Lockscreen when dreaming on device with SWIPE Security
When a device with a SWIPE Security started dreaming from the Gone scene, we have an extra transition to the Lockscreen while dreaming. This causes the dream transition to be unsmooth, and the device ends up in a state where the Dream and the Lockscreen scenes overlap.

The solution implemented adds the Lockscreen to the backstack, instead of switching the scene explicitly, and adds a state flow variable following backstack changes to correctly resolve to the Lockscreen in the HomeSceneFamilyResolver, if the lockscreen is on the backstack. The existing state flow variables did not capture this state change, as "isDeviceEnteredDirectly" filters out the Dream scene, and remains true, causing "isDeviceEntered" to also remain true. Since the device is always considered unlocked with SWIPE Security, and still entered, we cannot resolve to the Lockscreen in the HomeSceneFamilyResolver unless we add this extra state flow variable, or explicitly switch to the Lockscreen in this specific condition when the dream stops.

This CL also adds a test class for the HomeSceneFamilyResolver.

Flag: com.android.systemui.scene_container
Bug: 456071675
Test: flashed Blazer & TangorPro devices, and started dreaming while having a SWIPE security method
Test: atest HomeSceneFamilyResolverTest
Test: atest SceneBackInteractorTest
Test: atest DeviceEntryInteractorTest
Change-Id: I3e5fdf026f6421655d960437328ef9cc9d8b5412