Fixes relaunching pip app to bubble
Unable to start an app into a Bubble while the app is in PIP once
the bubble root task flag is enabled. That’s because the app task
failed to reparent to the bubble root task while it was in pip.
Reverting the change since it no longer work in PIP2.

In addition, bubble now update the override windowingMode on the
root task vs. the app task, but PIP updates the properties on the
app Task directly. So, the app task was still in pip windowingMode
while relaunching it as a bubble and was animated by PipHandler.

For task bounds, bringing back using ActivityOptions#setLaunchBounds
while starting the app to override the task bounds requested from PIP.
The leaf task override bounds would be clear along with updating the
bubble root task bounds later.

Bug: 462475923
Test: start an app into bubble from pip
Test: WindowContainerTransactionTests
Test: BubbleRootTaskTest
Flag: com.android.window.flags.enable_bubble_root_task
Change-Id: I0d098cf51deec6d1bcd18a783ed7eb8d92702023