[PiP on CD] Fix drag-PiP-across-display issues
Previously, config overrides were added to ActivityRecord for dragging
PiP across displays to block config propagation to the app to solve
b/422565128. However, this led to multiple issues with the density of
the PiP task in the new display in the following cases:
1) Drag PiP to a new display, then expanding PiP (see b/431280403)
2) Launch the task in PiP in a different display (see b/424569682)
Subsequent fixes were added but these led to even newer issues, including a frozen and unresponsive YouTube frame when PiP is expanded.

This CL removes the config overrides for dragging PiP across displays in
Core, and clean up the handling of launching the PiP task cross-display
and expanding after dragging PiP to a new display. Everything should now
be handled on Shell side.

Note that there are some remaining issues after this fix, notably with
the YouTube app pausing playback and disabling auto-enter PiP after the
task launch/expand in extended mode, but these issues are much less
minor and will be tracked and fixed in follow-up CLs. Manual testing
for this CL was done with YouTube and Netflix in both projected and
extended modes with different devices.

Bug: 449852735
Bug: 454496827
Test: atest DesktopPipTransitionControllerTests ActivityRecordTests
Test: atest WMShellUnitTests:com.android.wm.shell.pip2.phone
Test: Manual - PiP in display A, launch PiP task in display B - verify
task appears in display B with the correct density. Test vice versa
(swap the displays).
Test: Manual - PiP in display A, drag to display B and expand - verify
task appears in display B with the correct density. Test vice versa
(swap the displays).
Flag: EXEMPT BUGFIX

Change-Id: Iec9e119d94b89b1ed6ac18c4a6e21f6bf36fa828