Fix surface position of child window in a scaled parent
This is like to fix a typo in [1]. It was no problem if the child and
parent has the same scale (e.g. from the same app). But if the child
has a different scale, the position will be wrong, e.g. a parent
parent with 2x scale and child with 1x scale: A position (100,100)
of child will be scaled to (200,200) in screen coordinates. So if the
child requests to show at (100,100), the position should be (50,50).

The correct statement should be similar to [2], which inverses parent's
scale to set surface scale of child.

[1]: I6d4f7dab4949feb488fbcb42ff687a520fe4ca70
[2]: I83b32090ceb2e92d644e83b9b14ea68ef2ca96f2

Bug: 182362657
Flag: EXEMPT BUGFIX
Test: atest WindowStateTests#testCompatOverrideScale
Change-Id: I94a6dd63aaffb23fd641e09ab085e49a1f05251a