Fix fade choreography in Bouncer -> LS/Shade transition
The transition faded out first the background, then the content. This
 was mistakenly introduced during a previous refactoring (ag/33293230)

Additionally, this removes the triggered fade-in of the LS elements
when transitioning from bouncer. The spec is quite different, hence
landing a short term fix, and created b/464481844 for the remainder of
the work

The previous `contentAlphaAnimatable` trigger logic contained multiple
flaws
- `useFadeInOnComposition` was computed only during initial composition.
  This assumption does not hold for overlays (bouncer was one of them),
  and if a scene animation is interrupted before fully settling (this
  could be reproduced by fidgeting with the gesture)
- `currentTransition.isUserInputOngoing` was not a key to the
   `LaunchedEffect`, thus ending the gesture would not trigger the
   animation, only the end of the transition would.
Thus, replaces the logic with an implementation that is easier to read,
and addressed the aforementioned issues.

There is still an observable delay for fading in the lockscreen content
on slow gestured, however this is due to the multiple alpha values used
on `LockscreenElementKeys.Root` - the STL transition definition only
fades the content in after .8 of the transition, and cannot be changed
easily since the shade <-> lockscreen transition definition is currently
symmetric. This is for a follow-up CL

Bug: 457668950
Test: go/ag-37168311-manual-test
Flag: com.android.systemui.scene_container

Change-Id: I17d8eb303d0c0eaaf9e556985d08e7f8359ebc6c