LocalBluetoothAdapter to use context user as the default user
In the deprecated getInstance method LocalBluetoothAdapter was using hard coded UserHandle.SYSTEM which was incorrectly redirecting the call to the SYSTEM user in case of HSUM from the existing callers like Settings.

Changed the deprecated method to user context.getUser to use the current context user as the default user.

Bug: 463347191
Test: Tested the Pair new device flow is successfully listing bluetooth devices.
Flag: EXEMPT BUGFIX
Change-Id: I91357029512e4c289b7c3bcebd5b4fca659b55f2