Load system fonts in bindApplication for native services
The system server loads system fonts and passes the shared memory to
each (managed) application in `bindApplication()`. This improves
performance as an empty system font map results in a disk read on every
first Font NDK API invocation in the service process. This commit
implements this feature for native services.

Another reason for implementing this feature is to fix a bug -
currently, calling `AFontMatcher_match`, which is a Font NDK API, in a
native service crashes the process. This is because the call causes SEGV
when the font map is empty.

Test: Clank's native renderers work
Flag: android.os.native_framework_prototype
Bug: 463480321
Change-Id: Icf0367bb08513ddf01a0a23e46139de1fc955e51