Flexiglass: Disable notification scroll updates in QuickSettingsScene
The notification stack in `QuickSettingsScene` is used only during
transitions and should never be considered active in the QS context.

This change introduces an `isActivated` parameter to the notification
composable. When set to `false` — as is now done in
`QuickSettingsScene` — it prevents the reporting of `ShadeScrollState`

Test: Manually tested on SingleShade
Bug: 457767023
Flag: com.android.systemui.scene_container
Change-Id: I2c65d1040ef4a215af9e95e2d527f9f694dee359