Avoid NPE for a binder call
This change avoids possible NPE for a binder call of the IStatusBar.
The null check before usage is not enough to cover all the cases,
because during the null check and the actual call, the binder can die
and can become null when the StatusBarManagerService recelves the binder
die event as a DeathRecipient.

Introduce a new executor to run most of the binder calls to avoid
duplicated code everywhere and to reduce the risk of mis-use of the
non-thread-safe variable.

The change should reduce unintentional system server crash and improve
the stability of test results.

Calls with complex operations based on the results remains unchanged.

Bug: 462750237
Test: BumbleBluetoothTests
Flag: EXEMPT bugfix
Change-Id: I148f6a07e70409a960786904ee7bd7751b41eeab