Retain pipeline caches in memory permanently
In the original design of PipelineCache, with the assumption that the
cache would only be read once by Skia, we decided to drop the data
pointer as soon as the cache was read, to ensure that the file memory
mapping is released as soon as possible. In addition, we never updated
the cache data (or our bookkeeping) from writes as we assumed the cache
would never be loaded again.

However, Skia can read the cache multiple times - for example, if a
memory trimming command is sent to the app, Skia will drop and later
rebuild its graphics pipelines, reading the cache again.

When attempting to roll the flag forward, we ran into an issue where a
factory reset run of the calculator, followed by a memory trim, followed
by a relaunch of the calculator caused the PipelineCache to return an
empty cache. Since on the initial load there was no pipeline cache file,
we set `mHashCache` to false, never update it when caches are written,
and after the trim, we return an empty cache. As a result, Skia would
rebuild pipelines from scratch and we regressed several performance
metrics.

Since Skia might need the cache after a trim, and we want to avoid
blocking the cache load hotpath with system calls and disk I/O, we can
resolve the issue by always keeping the latest cache in memory.

Note that this also significantly simplifies the code.

Alternative solutions are to always reload the cache, or to explicitly
memory map the cache after a trim, in preparation for resuming the app
(although this may result in unexpected I/O activity when system
resources are scarce).

Bug: 269117286
Test: boot
Test: adb shell am send-trim-memory com.google.android.calculator 80
Test: inspect perfetto trace
Test: atest hwui_unit_tests -- --test-arg com.android.tradefed.testtype.GTest:native-test-flag:"--renderer=skiavk"
Flag: com.android.graphics.hwui.flags.separate_pipeline_cache
Change-Id: I6a9ffe64740385e5e5687e5eb802ff952228b02f