Add WebApp service name to Context API
7341ffc853574b6805be9be74e8b678f801d7a4b broke many builds due to the
inclusion of WebApp config flags, which contains an exported flag. The
exported flag is used to guard new APIs added in the WebApp module,
however it's unused on the platform, causing "exported but not used to
guard any API" errors.

To fix this, let's add an API in `Context` class for the service name
and remove the one inside the module itself later.

Bug: 464781604
Bug: 432147348
Flag: com.android.webapp.flags.enable_web_app_service
Test: build panther-trunk_food-userdebug
Change-Id: Ibb646330df94b0f31fd2a8e874ca0531f0461939