Set alwaysOnTop earlier for split-to-bubble transition
With bubble root task enabled, the Bubble module sets the alwaysOnTop
property to the bubble root task in BubbleTaskViewListener#onTaskCreated
vs to each bubble task in LaunchOrConvertToBubble#onInflated.
Then if a split-to-bubble transition happens, the SplitHandler will
reparent the remaining task to the top of TDA, which occludes the bubble
root task and the Bubble root task hasn't been always-on-top, which
triggers the transition to move the bubble root task to back and
LaunchOrConvertToBubble can't handle the transition properly.

This CL sets the always-on-top property when handleRequest.

Test: manual - try the split-tobubble transition
Test: atest RelaunchSplitScreenToBubbleTest
Bug: 461693176
Flag: EXEMPT bugfix

Change-Id: I3f843861f4777319fbc29f81787c253a385b2380