Enable DEPRECATE_USES_CLEARTEXT_TRAFFIC compat change after Cinnamon Bun.
This was always the plan, but I thought @EnableAfter was inclusive, and that I had to wait for the new version to be in the Enum.
It's not, so this change will be enabled for the release following 26Q2 as it should be.

Bug: 415007211
Flag: android.security.deprecate_uses_cleartext_traffic2
Change-Id: I7c6b22257ace9a31032833ba5ceef9ee7c2a5845