Automatically initialize ProtoLog async on first log call if not done yet
This ensure we never run into the case of ProtoLog messages missing
because they are logged before ProtoLog is initialized.

The caveat is that this will add some additional to the first ProtoLog
call if we haven't called ProtoLog.init in the process earlier, so we
may want to think of ways to optimize this further in the future.

Flag: android.tracing.protolog_async_init
Bug: 464952546
Test: atest TracingTests
Change-Id: Ic63362966f28edcc5ed7d0523edb72a50c6c70bd