Trigger throttled auth warm-up on bouncer user input.
Trigger a warm-up signal to LockPatternUtils when the user provides input on the primary bouncer.
The `AuthenticationInteractor` throttles to a max of one per 5s, allowing the first trigger through.
The warm-up is triggered by user input events in `KeyguardSecurityContainerController` and `BouncerInteractor`,
to cover both Flexiglass and non-Flexiglass.

Bug: 455600390
Test: Manual on flexiglass on and off
Test: atest SystemUITests
Flag: android.security.enable_weaver_warmup
Change-Id: I26de52c0a8193b127fcf4881424263e6798afa99