large screen capture: save selected capture type and region
Add methods in LargeScreenCaptureParametersRepository to save and
read the user selected capture type and region in user secure
settings. When a PreCaptureModelView is activated, it will first
check if the given screen capture parameter has default type and
region (e.g. when the UI is started by the keyboard shortcuts).
If it does, then use the default parameters as the initial type
and region and save them. Otherwise, use the previous selected
type and region as the initial value.

Demos are in the bug comment.

Bug: 423697394
Test: Manual, PreCaptureViewModelTest, LargeScreenCaptureParametersInteractor/RepositoryTest
Flag: com.android.systemui.large_screen_recording
Change-Id: I8d118783c7cf1dcc160772129591d44e2b908240