Move x86_cpu_energy_attribution flag to backstage_power
The desktop_pnp namespace is a Desktop namespace. This flag needs to be
in a platform-level namespace to function correctly in the Desktop
release environment because the code using this flag resides in
frameworks/base, which is outside vendor/google/desktop.

Bug: 465479558
Test: Build
Flag: com.android.server.power.optimization.x86_cpu_energy_attribution
Change-Id: I653e01c7aff0cbf1278fc74e09a5ceb9b70cb7a5