Refactor USB grantPermission apis
In preparation for an implementation that allows persisting USB
permissions, refactor the grantPermission apis for UsbDevice and
UsbAccessory to also provide packageName and a persistence flag. These
fields are not used yet.

The major change is that all of the grantPermission apis that only
provide uid to also provide packageName. Per
PackageManager#getPackagesForUid, you can't get a 1-1 mapping of uid to
package name and this will be necessary for adding apis to revoke
permissions. While the api is changed to pass the packageName now in all
APIs, it won't be used until the accompanying flag is enabled (in
a subsequent commit).

This change also removes existing apis for persisting
permissions in IUsbManager.aidl. They are currently unused and conflict
with the refactored grantPermission api.

Bug: 389062495
Test: atest UsbTests
Test: Build and deploy to device
Flag: EXEMPT PURE_REFACTOR
Change-Id: I257638dd13b113cc31f8a8a74b71c592fc8f68fd