Remove unused config_skinnyNotifsInLandscape
This was previously used to tell notifications whether they should take
up the whole available horizontal space or be limited to the width of
two QS tiles. It was essentially true when we were showing the split
shade, so over time it ended up not being used at all aside from some
logging, since we can directly check whether we're in split shade
instead.

I also removed mIsSmallLandscapeLockscreenEnabled which was associated
with a cut feature (landscape keyguard), and the fields that were being
logged for it.

Bug: 433921584
Test: presubmit
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ia7cb30ec5f46f70f8bb6e99f0fc6bff57e5a52e7