Only animate action when attached
When notificaitons are inflated they come from a different thread,
starting an animator can cause a race condition where messages on
MessageQueue will forever get stuck.

By deferring to the view being attached, we'll guarantee that
this is called on the correct thread by making it a side-effect of a UI
toolkit callback.

Bug: 464275874
Flag: EXEMPT Bugfix
Test: manual / QA team
Test: patch ag/37290366 and use `adb shell cmd notification notify` to post a notification that generates actions (see b/464275874 for details)

Change-Id: Ib5191a98e74373af6b76a0123083bab23a2b8607