ComputerControl: prevent agents from closing the last task via BACK
Agentic journeys often fail because an agent mistakenly closes an
app when they mispredict what back is going to do, e.g. expecting
it to close a cling but then the entire app closes. To avoid that,
we instead move the task to back on computer control displays.

Note that if there is only a single task on the computer control
display, this will be a no-op.

Instead of simply dropping the action, we may want to report back
to the agent that it would have closed the app in a future CL.

Bug: 459913303
Test: atest 'ActivityRecordTests#testBackOnTaskRoot_onComputerControlDisplay_movesToBack'
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: Ibaf9d16c65afa6e045f7d81554f5854dc757e532