Validate activity launch success for app sharing
This change introduces a check on the WaitResult provided by the asyncActivityLauncher.startActivityAsUser callback. If the activity launch result is neither ActivityManager.START_SUCCESS nor ActivityManager.START_TASK_TO_FRONT, the media projection is explicitly denied by setting the sharing state to Denied. This ensures that app sharing is only initiated when the target application is confirmed to be in a shareable state.

Bug: 466443910
Test: manual
Flag: com.android.systemui.large_screen_sharing
Change-Id: I29adb85682257d54a51eebf3a81625794c704eaf